Removed LevelDB support, fixed the HTTP request function and the cancellation
Removed LevelDB support for reading the local storage of Discord. Discord now uses snappy compression and the library doesn't seem to support it yet so it's not worth the effort to fix it. Also modified the HTTP request function to try again on 500 error, this might fix the problem where some messages were not deleted.